+++ components/src/dynamo/vllm/omni/audio_handler.py
@@ -7,22 +7,13 @@ Extracted from omni_handler.py to keep modality-specific logic separate.
 OmniHandler holds an instance as ``self.audio`` (composition).
 """
 
-import asyncio
 import base64
 import logging
-import time
-import uuid
-from io import BytesIO
 from typing import Any, Dict
 
 from vllm_omni.inputs.data import OmniTextPrompt
 
-from dynamo.common.protocols.audio_protocol import (
-    AudioData,
-    NvAudioSpeechResponse,
-    NvCreateAudioSpeechRequest,
-)
-from dynamo.common.storage import upload_to_fs
+from dynamo.common.protocols.audio_protocol import NvCreateAudioSpeechRequest
 from dynamo.common.utils.output_modalities import RequestType
 
 logger = logging.getLogger(__name__)
@@ -426,157 +417,3 @@ class AudioGenerationHandler:
                 "Failed to estimate TTS prompt length, using fallback 2048: %s", e
             )
             return 2048
-
-    # -- Audio output formatting ----------------------------------------------
-
-    def _extract_audio_tensor(self, mm_output: Dict[str, Any]) -> tuple:
-        """Extract audio tensor and sample rate from multimodal_output dict."""
-        import numpy as np
-        import torch
-
-        audio_key = "audio" if "audio" in mm_output else "model_outputs"
-        audio_val = mm_output.get(audio_key)
-        if audio_val is None:
-            raise ValueError(
-                f"No audio data in multimodal_output. Keys: {list(mm_output.keys())}"
-            )
-
-        if isinstance(audio_val, list):
-            audio_val = torch.cat(audio_val, dim=-1)
-
-        if hasattr(audio_val, "float"):
-            audio_np = audio_val.float().detach().cpu().numpy()
-        elif isinstance(audio_val, np.ndarray):
-            audio_np = audio_val.astype(np.float32)
-        else:
-            audio_np = np.array(audio_val, dtype=np.float32)
-
-        if audio_np.ndim > 1:
-            audio_np = audio_np.squeeze()
-
-        sr_raw = mm_output.get("sr", 24000)
-        if isinstance(sr_raw, list):
-            sr_raw = sr_raw[-1] if sr_raw else 24000
-        sample_rate = sr_raw.item() if hasattr(sr_raw, "item") else int(sr_raw)
-
-        return audio_np, sample_rate
-
-    def _encode_audio(
-        self, audio_np, sample_rate: int, fmt: str = "wav", speed: float = 1.0
-    ) -> tuple:
-        """Encode a numpy float32 waveform to audio bytes.
-
-        Uses soundfile for multi-format support.
-        Applies speed adjustment via librosa if speed != 1.0.
-        """
-        import soundfile as sf
-
-        if speed != 1.0:
-            try:
-                import librosa
-
-                audio_np = librosa.effects.time_stretch(y=audio_np, rate=speed)
-            except ImportError:
-                logger.warning("librosa not installed, ignoring speed adjustment")
-
-        fmt = (fmt or "wav").lower()
-        format_map = {
-            "wav": ("WAV", "audio/wav", {}),
-            "pcm": ("RAW", "audio/pcm", {"subtype": "PCM_16"}),
-            "flac": ("FLAC", "audio/flac", {}),
-            "mp3": ("MP3", "audio/mpeg", {}),
-            "aac": ("AAC", "audio/aac", {}),
-            "opus": ("OGG", "audio/ogg", {"subtype": "OPUS"}),
-        }
-
-        if fmt not in format_map:
-            logger.warning(f"Unsupported format '{fmt}', defaulting to wav")
-            fmt = "wav"
-
-        sf_format, media_type, kwargs = format_map[fmt]
-
-        buf = BytesIO()
-        sf.write(buf, audio_np, sample_rate, format=sf_format, **kwargs)
-        return buf.getvalue(), media_type
-
-    async def format_output(
-        self,
-        mm_output: Dict[str, Any],
-        request_id: str,
-        response_format: str | None = None,
-        request_type: RequestType = RequestType.AUDIO_GENERATION,
-        speed: float = 1.0,
-    ) -> Dict[str, Any] | None:
-        """Format multimodal audio output for the response."""
-        if not mm_output:
-            return NvAudioSpeechResponse(
-                id=request_id,
-                model=self.config.served_model_name or self.config.model,
-                status="failed",
-                created=int(time.time()),
-                error="No audio generated",
-            ).model_dump()
-
-        try:
-            start_time = time.time()
-
-            audio_np, sample_rate = self._extract_audio_tensor(mm_output)
-
-            encode_fmt = (
-                "wav"
-                if response_format in (None, "url", "b64_json")
-                else response_format
-            )
-
-            assert encode_fmt is not None
-            audio_bytes, media_type = await asyncio.to_thread(
-                self._encode_audio, audio_np, sample_rate, encode_fmt, speed
-            )
-
-            logger.info(
-                f"Audio encoded for request {request_id}: "
-                f"{len(audio_np)} samples, sr={sample_rate}, "
-                f"{len(audio_bytes)} bytes {encode_fmt}"
-            )
-
-            inference_time = time.time() - start_time
-
-            if response_format == "url":
-                ext = encode_fmt if encode_fmt != "opus" else "ogg"
-                storage_path = f"audios/{request_id}/{uuid.uuid4()}.{ext}"
-                url = await upload_to_fs(
-                    self.media_output_fs,
-                    storage_path,
-                    audio_bytes,
-                    self.media_output_http_url,
-                )
-                audio_data_obj = AudioData(url=url)
-            else:
-                b64 = base64.b64encode(audio_bytes).decode("utf-8")
-                audio_data_obj = AudioData(b64_json=b64)
-
-            response = NvAudioSpeechResponse(
-                id=request_id,
-                object="audio.speech",
-                model=self.config.served_model_name or self.config.model,
-                status="completed",
-                progress=100,
-                created=int(time.time()),
-                data=[audio_data_obj],
-                inference_time_s=inference_time,
-            )
-            return response.model_dump()
-
-        except Exception as e:
-            logger.error(f"Failed to process audio for request {request_id}: {e}")
-            error_response = NvAudioSpeechResponse(
-                id=request_id,
-                object="audio.speech",
-                model=self.config.served_model_name or self.config.model,
-                status="failed",
-                progress=0,
-                created=int(time.time()),
-                data=[],
-                error=str(e),
-            )
-            return error_response.model_dump()
