+++ components/src/dynamo/vllm/tests/omni/test_audio_handler.py
@@ -190,94 +190,3 @@ class TestEngineInputsFromAudio:
         req = NvCreateAudioSpeechRequest(input="hello", speed=2.0)
         inputs = await handler.build_engine_inputs(req)
         assert inputs.speed == 2.0
-
-
-class TestExtractAudioTensor:
-    """Tests for _extract_audio_tensor."""
-
-    def test_extracts_from_audio_key(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        mm = {"audio": np.array([0.1, -0.2, 0.3], dtype=np.float32), "sr": 24000}
-        audio_np, sr = handler._extract_audio_tensor(mm)
-        assert sr == 24000
-        assert len(audio_np) == 3
-
-    def test_extracts_from_model_outputs_key(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        mm = {
-            "model_outputs": np.array([0.5, -0.5], dtype=np.float32),
-            "sr": 16000,
-        }
-        audio_np, sr = handler._extract_audio_tensor(mm)
-        assert sr == 16000
-        assert len(audio_np) == 2
-
-    def test_missing_audio_raises(self):
-        handler = _make_audio_handler()
-        with pytest.raises(ValueError, match="No audio data"):
-            handler._extract_audio_tensor({"sr": 24000})
-
-    def test_squeezes_extra_dims(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        mm = {"audio": np.array([[0.1, 0.2, 0.3]], dtype=np.float32), "sr": 24000}
-        audio_np, _ = handler._extract_audio_tensor(mm)
-        assert audio_np.ndim == 1
-
-
-class TestEncodeAudio:
-    """Tests for _encode_audio."""
-
-    def test_wav_encoding(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        audio = np.zeros(2400, dtype=np.float32)
-        audio_bytes, media_type = handler._encode_audio(audio, 24000, "wav")
-        assert media_type == "audio/wav"
-        assert len(audio_bytes) > 0
-        assert audio_bytes[:4] == b"RIFF"  # WAV header
-
-    def test_unsupported_format_falls_back_to_wav(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        audio = np.zeros(100, dtype=np.float32)
-        _, media_type = handler._encode_audio(audio, 24000, "xyz")
-        assert media_type == "audio/wav"
-
-    def test_default_format_is_wav(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        audio = np.zeros(100, dtype=np.float32)
-        _, media_type = handler._encode_audio(audio, 24000)
-        assert media_type == "audio/wav"
-
-
-class TestFormatAudioChunk:
-    """Tests for format_output."""
-
-    @pytest.mark.asyncio
-    async def test_empty_mm_output_returns_error(self):
-        handler = _make_audio_handler()
-        result = await handler.format_output({}, "req-1")
-        assert result["status"] == "failed"
-        assert "No audio generated" in result["error"]
-
-    @pytest.mark.asyncio
-    async def test_successful_generation(self):
-        import numpy as np
-
-        handler = _make_audio_handler()
-        mm = {"audio": np.random.randn(4800).astype(np.float32), "sr": 24000}
-        result = await handler.format_output(mm, "req-1")
-        assert result["status"] == "completed"
-        assert result["object"] == "audio.speech"
-        assert len(result["data"]) == 1
-        assert result["data"][0]["b64_json"] is not None
